+++ src/components/load-button.js
@@ -1,17 +1,6 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from '../components/abstract-component.js';
 
-export default class ButtonLoadMore {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
+export default class ButtonLoadMore extends AbstractComponent {
 
   getTemplate() {
     return `<button class="load-more" type="button">load more</button>`;
