+++ src/components/menu.js
@@ -1,17 +1,6 @@
-import {createElement} from '../utils.js';
+import AbstractComponent from '../components/abstract-component.js';
 
-export default class Menu {
-  constructor() {
-    this._element = null;
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-
-    return this._element;
-  }
+export default class Menu extends AbstractComponent {
 
   getTemplate() {
     return `<section class="control__btn-wrap">
